Move locality checker to separate subproject
# Conflicts:
#	formver.compiler-plugin/build.gradle.kts
#	formver.compiler-plugin/plugin/src/org/jetbrains/kotlin/formver/plugin/compiler/PluginErrors.kt

# Conflicts:
#	formver.compiler-plugin/plugin/src/org/jetbrains/kotlin/formver/plugin/compiler/PluginErrors.kt
