+++ formver.compiler-plugin/analysis/build.gradle.kts
@@ -4,9 +4,6 @@ plugins {
 
 dependencies {
     compileOnly(kotlin("compiler"))
-    implementation(project(":formver.common"))
-    implementation(project(":formver.compiler-plugin:plugin"))
-    implementation(project(":formver.compiler-plugin:locality"))
 }
 
 sourceSets {
