+++ formver.compiler-plugin/locality/src/org/jetbrains/kotlin/formver/locality/plugin/LocalityErrorMessages.kt
@@ -1,0 +1,23 @@
+/*
+ * Copyright 2010-2026 JetBrains s.r.o. and Kotlin Programming Language contributors.
+ * Use of this source code is governed by the Apache 2.0 license that can be found in the license/LICENSE.txt file.
+ */
+
+package org.jetbrains.kotlin.formver.locality.plugin
+
+import org.jetbrains.kotlin.diagnostics.KtDiagnosticFactoryToRendererMap
+import org.jetbrains.kotlin.diagnostics.rendering.BaseDiagnosticRendererFactory
+import org.jetbrains.kotlin.diagnostics.rendering.CommonRenderers
+import org.jetbrains.kotlin.formver.locality.plugin.extension.LocalityRenderer
+
+object LocalityErrorMessages : BaseDiagnosticRendererFactory() {
+    override val MAP: KtDiagnosticFactoryToRendererMap by KtDiagnosticFactoryToRendererMap("FormalVerificationLocality") { map ->
+        map.put(
+            LocalityErrors.LOCALITY_VIOLATION,
+            "{0} locality mismatch: expected {1}, actual {2}.",
+            CommonRenderers.STRING,
+            LocalityRenderer,
+            LocalityRenderer,
+        )
+    }
+}
