+++ formver.compiler-plugin/locality/src/org/jetbrains/kotlin/formver/locality/plugin/LocalityExtensionRegistrar.kt
@@ -3,12 +3,14 @@
  * Use of this source code is governed by the Apache 2.0 license that can be found in the license/LICENSE.txt file.
  */
 
-package org.jetbrains.kotlin.formver.plugin.compiler
+package org.jetbrains.kotlin.formver.locality.plugin
 
 import org.jetbrains.kotlin.fir.extensions.FirExtensionRegistrar
+import org.jetbrains.kotlin.formver.locality.plugin.extension.LocalityAdditionalCheckers
 
 object LocalityExtensionRegistrar : FirExtensionRegistrar() {
     override fun ExtensionRegistrarContext.configurePlugin() {
+        registerDiagnosticContainers(LocalityErrors)
         +LocalityAdditionalCheckers.getFactory()
     }
 }
