+++ formver.compiler-plugin/locality/src/org/jetbrains/kotlin/formver/locality/plugin/extension/LocalityPropertyChecker.kt
@@ -3,29 +3,29 @@
  * Use of this source code is governed by the Apache 2.0 license that can be found in the license/LICENSE.txt file.
  */
 
-package org.jetbrains.kotlin.formver.plugin.compiler.locality
+package org.jetbrains.kotlin.formver.locality.plugin.extension
 
 import org.jetbrains.kotlin.diagnostics.DiagnosticReporter
 import org.jetbrains.kotlin.diagnostics.reportOn
 import org.jetbrains.kotlin.fir.analysis.checkers.MppCheckerKind
 import org.jetbrains.kotlin.fir.analysis.checkers.context.CheckerContext
-import org.jetbrains.kotlin.fir.analysis.checkers.declaration.FirValueParameterChecker
-import org.jetbrains.kotlin.fir.declarations.FirValueParameter
-import org.jetbrains.kotlin.formver.plugin.compiler.PluginErrors.LOCALITY_VIOLATION
+import org.jetbrains.kotlin.fir.analysis.checkers.declaration.FirPropertyChecker
+import org.jetbrains.kotlin.fir.declarations.FirProperty
+import org.jetbrains.kotlin.formver.locality.plugin.LocalityErrors.LOCALITY_VIOLATION
 
-object LocalityValueParameterChecker : FirValueParameterChecker(MppCheckerKind.Common) {
+object LocalityPropertyChecker : FirPropertyChecker(MppCheckerKind.Common) {
     context(context: CheckerContext, reporter: DiagnosticReporter)
-    override fun check(declaration: FirValueParameter) {
-        val defaultValue = declaration.defaultValue ?: return
-        val requiredLocality = declaration.extractRequiredLocality()
-        val actualLocality = defaultValue.extractLocality()
+    override fun check(declaration: FirProperty) {
+        val initializer = declaration.initializer ?: return
+        val requiredLocality = declaration.extractLocality()
+        val actualLocality = initializer.extractLocality()
 
         if (requiredLocality.accepts(actualLocality)) return
 
         reporter.reportOn(
-            defaultValue.source ?: declaration.source,
+            initializer.source,
             LOCALITY_VIOLATION,
-            "Initializer",
+            "Assignment",
             requiredLocality,
             actualLocality
         )
