+++ formver.compiler-plugin/locality/src/org/jetbrains/kotlin/formver/locality/plugin/extension/LocalityVariableAssignmentChecker.kt
@@ -3,27 +3,26 @@
  * Use of this source code is governed by the Apache 2.0 license that can be found in the license/LICENSE.txt file.
  */
 
-package org.jetbrains.kotlin.formver.plugin.compiler.locality
+package org.jetbrains.kotlin.formver.locality.plugin.extension
 
 import org.jetbrains.kotlin.diagnostics.DiagnosticReporter
 import org.jetbrains.kotlin.diagnostics.reportOn
 import org.jetbrains.kotlin.fir.analysis.checkers.MppCheckerKind
 import org.jetbrains.kotlin.fir.analysis.checkers.context.CheckerContext
-import org.jetbrains.kotlin.fir.analysis.checkers.declaration.FirPropertyChecker
-import org.jetbrains.kotlin.fir.declarations.FirProperty
-import org.jetbrains.kotlin.formver.plugin.compiler.PluginErrors.LOCALITY_VIOLATION
+import org.jetbrains.kotlin.fir.analysis.checkers.expression.FirVariableAssignmentChecker
+import org.jetbrains.kotlin.fir.expressions.FirVariableAssignment
+import org.jetbrains.kotlin.formver.locality.plugin.LocalityErrors.LOCALITY_VIOLATION
 
-object LocalityPropertyChecker : FirPropertyChecker(MppCheckerKind.Common) {
+object LocalityVariableAssignmentChecker : FirVariableAssignmentChecker(MppCheckerKind.Common) {
     context(context: CheckerContext, reporter: DiagnosticReporter)
-    override fun check(declaration: FirProperty) {
-        val initializer = declaration.initializer ?: return
-        val requiredLocality = declaration.extractLocality()
-        val actualLocality = initializer.extractLocality()
+    override fun check(expression: FirVariableAssignment) {
+        val requiredLocality = expression.lValue.extractLocality()
+        val actualLocality = expression.rValue.extractLocality()
 
         if (requiredLocality.accepts(actualLocality)) return
 
         reporter.reportOn(
-            initializer.source,
+            expression.rValue.source,
             LOCALITY_VIOLATION,
             "Assignment",
             requiredLocality,
