+++ formver.compiler-plugin/plugin/src/org/jetbrains/kotlin/formver/plugin/compiler/FormalVerificationPluginErrorMessages.kt
@@ -8,7 +8,6 @@ package org.jetbrains.kotlin.formver.plugin.compiler
 import org.jetbrains.kotlin.diagnostics.KtDiagnosticFactoryToRendererMap
 import org.jetbrains.kotlin.diagnostics.rendering.BaseDiagnosticRendererFactory
 import org.jetbrains.kotlin.diagnostics.rendering.CommonRenderers
-import org.jetbrains.kotlin.formver.plugin.compiler.locality.LocalityRenderer
 
 object FormalVerificationPluginErrorMessages : BaseDiagnosticRendererFactory() {
     override val MAP: KtDiagnosticFactoryToRendererMap by KtDiagnosticFactoryToRendererMap("FormalVerification") { map ->
@@ -66,13 +65,6 @@ object FormalVerificationPluginErrorMessages : BaseDiagnosticRendererFactory() {
             PluginErrors.PURITY_VIOLATION,
             "{0}",
             CommonRenderers.STRING,
-        )
-        map.put(
-            PluginErrors.LOCALITY_VIOLATION,
-            "{0} locality mismatch: expected {1}, actual {2}.",
-            CommonRenderers.STRING,
-            LocalityRenderer,
-            LocalityRenderer,
         )
         map.put(
             PluginErrors.UNIQUENESS_VIOLATION,
