+++ settings.gradle.kts
@@ -26,4 +26,6 @@ include("formver.compiler-plugin:cli")
 include("formver.compiler-plugin:uniqueness")
 include("formver.compiler-plugin:viper")
 include("formver.compiler-plugin:plugin")
+include("formver.compiler-plugin:locality")
+include("formver.compiler-plugin:analysis")
 include("formver.compiler-plugin:core")
